+++ JAVA/MundoPc/MundoPc/nbproject/private/private.properties
@@ -1,0 +1,1 @@
+user.properties.file=C:\\Users\\gabriel\\AppData\\Roaming\\NetBeans\\12.1\\build.properties
