+++ JAVA/MundoPc/MundoPc/nbproject/private/private.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
+    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
+        <group>
+            <file>file:/C:/Users/gabriel/Desktop/T/Cursos2023/UTN/Tecnicatura3Repo/Team-Fortran-2023/MundoPc/src/mundoPc/MundoPc.java</file>
+        </group>
+    </open-files>
+</project-private>
