+++ PYTHON/Catalogo_pelicula/Catalogo_pelicula/.idea/.gitignore
