+++ PYTHON/Catalogo_pelicula/Catalogo_pelicula/.idea/inspectionProfiles/profiles_settings.xml
