+++ PYTHON/Catalogo_pelicula/Catalogo_pelicula/Dominio/pelicula.py
